feat(hw/gpu): added linux support
